refactor: apply DRY concept
